+++ packages/screenshot-utils/helpers.ts
@@ -44,12 +44,10 @@ export const customSnapshotIdentifier = ({
     return kebab(`${currentTestName}${counter > 1 ? `-${counter}` : ''}`);
 };
 
-const getPageHtml = async (page: Page, css?: string, theme?: string) => {
+const getPageHtml = async (page: Page, css?: string) => {
     const [head, body] = await Promise.all([page?.innerHTML('head'), page?.innerHTML('body')]);
 
-    const themeAttr = theme ? ` data-theme="${theme}"` : '';
-
-    return `<html><head>${head}</head><body${themeAttr}><style>${css}</style>${body}</body></html>`;
+    return `<html><head><style>${css}</style>${head}</head><body>${body}</body></html>`;
 };
 
 export type MatchHtmlParams = {
@@ -59,7 +57,6 @@ export type MatchHtmlParams = {
     matchImageSnapshotOptions?: MatchImageSnapshotOptions;
     screenshotOpts?: ScreenshotOpts;
     evaluate?: EvaluateFn;
-    theme?: string;
     viewport?: { width: number; height: number };
 };
 
@@ -81,10 +78,9 @@ export const matchHtml = async ({
     matchImageSnapshotOptions,
     screenshotOpts = screenshotDefaultOpts,
     evaluate,
-    theme,
     viewport = defaultViewport,
 }: MatchHtmlParams) => {
-    const pageHtml = await getPageHtml(page, css, theme);
+    const pageHtml = await getPageHtml(page, css);
 
     const image = await axios.post(
         playwrightUrl,