+++ packages/select/src/typings.ts
@@ -8,6 +8,7 @@ import {
     ReactElement,
 } from 'react';
 import { PopoverProps } from '@alfalab/core-components-popover';
+import { InputProps } from '@alfalab/core-components-input';
 
 export type OptionShape = {
     /**
@@ -139,7 +140,7 @@ export type BaseSelectProps = {
     /**
      * Подсказка под полем
      */
-    hint?: string;
+    hint?: ReactNode;
 
     /**
      * Возможность использовать селект как input-autocomplete
@@ -338,7 +339,7 @@ export type FieldProps = {
     /**
      * Подсказка под полем
      */
-    hint?: string;
+    hint?: ReactNode;
 
     /**
      * Компонент стрелки
@@ -440,6 +441,11 @@ export type OptionsListProps = {
      * Идентификатор для систем автоматизированного тестирования
      */
     dataTestId?: string;
+
+    /**
+     * Дополнительные пропсы для Input'a, находящегося внутри кастомного OptionsList
+     */
+    inputProps?: InputProps;
 };
 
 export type OptgroupProps = {