+++ packages/tabs/src/component.screenshots.test.tsx
@@ -31,7 +31,7 @@ describe('Tabs | TabsDesktop', () => {
             matchImageSnapshotOptions: {
                 failureThresholdType: 'pixel',
                 // TODO: ширина линии на сервере чуть больше
-                failureThreshold: 20,
+                failureThreshold: 40,
                 customSnapshotIdentifier: (...args) =>
                     `${theme}-${customSnapshotIdentifier(...args)}`,
             },