+++ packages/bank-card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-bank-card@4.0.4...@alfalab/core-components-bank-card@4.0.5) (2021-08-11)
+
+**Note:** Version bump only for package @alfalab/core-components-bank-card
+
+
+
+
+
+## [4.0.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-bank-card@4.0.3...@alfalab/core-components-bank-card@4.0.4) (2021-08-04)
+
+**Note:** Version bump only for package @alfalab/core-components-bank-card
+
+
+
+
+
 ## [4.0.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-bank-card@4.0.2...@alfalab/core-components-bank-card@4.0.3) (2021-07-23)
 
 **Note:** Version bump only for package @alfalab/core-components-bank-card